fix shared search redirect
